# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Risk & Disputes API
  version: 0.2.0
  description: Risk engine, fraud signals, dispute management, evidence, and representment
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3003
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
  /api/v1/disputes:
    get:
      summary: List disputes
      description: Returns a paginated list of disputes. Optionally filter by status or payment_intent_id.
      x-platform:
        kind: resource
        entity: Dispute
        affectsLedger: false
      operationId: listDisputes
      tags:
        - Disputes
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            type: string
            enum:
              - needs_response
              - under_review
              - won
              - lost
              - accepted
              - expired
        - name: payment_intent_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of disputes.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Dispute"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create dispute
      description: Creates a dispute against a payment intent.
      x-platform:
        kind: resource
        entity: Dispute
        emitsEvents:
          - dispute.created
        affectsLedger: false
      operationId: createDispute
      tags:
        - Disputes
      security:
        - MerchantKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payment_intent_id
                - reason
                - amount
                - currency
              properties:
                payment_intent_id:
                  type: string
                reason:
                  type: string
                  enum:
                    - fraudulent
                    - duplicate
                    - not_received
                    - product_unacceptable
                    - subscription_canceled
                    - other
                amount:
                  type: integer
                  minimum: 1
                currency:
                  type: string
                  minLength: 3
                  maxLength: 3
                due_by:
                  type: string
                  format: date-time
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "201":
          description: Dispute created.
          content:
            application/json:
              schema:
                type: object
                required:
                  - dispute
                properties:
                  dispute:
                    $ref: "#/components/schemas/Dispute"
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment intent not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/disputes/{id}:
    get:
      summary: Get dispute by ID
      description: Returns a single dispute.
      x-platform:
        kind: resource
        entity: Dispute
        affectsLedger: false
      operationId: getDisputeById
      tags:
        - Disputes
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Dispute found.
          content:
            application/json:
              schema:
                type: object
                required:
                  - dispute
                properties:
                  dispute:
                    $ref: "#/components/schemas/Dispute"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Dispute not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update dispute
      description: Updates metadata or due_by on an open dispute. Cannot update closed disputes.
      x-platform:
        kind: transition
        entity: Dispute
        emitsEvents:
          - dispute.updated
        affectsLedger: false
        stateChanges:
          - metadata_updated
      operationId: updateDispute
      tags:
        - Disputes
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  type: object
                  additionalProperties: true
                due_by:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Dispute updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - dispute
                properties:
                  dispute:
                    $ref: "#/components/schemas/Dispute"
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Dispute not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Dispute is closed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/risk/decisions/{id}:
    get:
      summary: Get risk decision by ID
      description: Returns a single risk decision.
      x-platform:
        kind: resource
        entity: RiskDecision
        affectsLedger: false
      operationId: getRiskDecisionById
      tags:
        - RiskDecisions
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Risk decision found.
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Risk decision not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update risk decision
      description: Updates an existing risk decision (e.g. override outcome).
      x-platform:
        kind: transition
        entity: RiskDecision
        emitsEvents:
          - risk_decision.updated
        affectsLedger: false
        stateChanges:
          - pending -> reviewed
          - metadata_updated
      operationId: updateRiskDecision
      tags:
        - RiskDecisions
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome:
                  type: string
                  enum:
                    - approve
                    - decline
                    - review
                reason:
                  type: string
      responses:
        "200":
          description: Risk decision updated.
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Risk decision not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
    Dispute:
      description: A payment dispute.
      type: object
      properties:
        id:
          type: string
        merchant_id:
          type: string
        payment_intent_id:
          type: string
        status:
          type: string
          enum:
            - needs_response
            - under_review
            - won
            - lost
            - accepted
            - expired
        reason:
          type: string
          enum:
            - fraudulent
            - duplicate
            - not_received
            - product_unacceptable
            - subscription_canceled
            - other
        amount:
          type: integer
        currency:
          type: string
        due_by:
          type: string
          format: date-time
          nullable: true
        closed_at:
          type: string
          format: date-time
          nullable: true
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...